backend/requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
fastapi==0.89.1
ffmpeg-python==0.2.0
filelock==3.9.0
future==0.18.3
h11==0.14.0
httptools==0.5.0
huggingface-hub==0.12.0
idna==3.4
more-itertools==9.0.0
numpy==1.24.2
openai-whisper==20230124
packaging==23.0
pydantic==1.10.4
pydub==0.25.1
python-dotenv==0.21.1
python-multipart==0.0.5
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tokenizers==0.13.2
torch==1.13.1
tqdm==4.64.1
transformers==4.26.0
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
